Adds form handling for CreateNewFolder modal.
Closes department-of-veterans-affairs/messaging-fe#68

- Renamed handleSubmit method, reordered validateFolderName parameters.
- Moved form validation back to ModalCreateFolder.
- Handles create new folder response by appending response to data.items.
- Closes Create New Folder modal onSubmit.

Unblocks department-of-veterans-affairs/messaging-fe#82
